akevisionback/requirements.txt
asgiref
certifi
chardet
Django
django-cors-headers
django-filter
djangorestframework
httpie
idna
Pygments
PySocks
pytz
requests
requests-toolbelt
sqlparse
urllib3
pyexcel-io
pyexcel-xlsx
channels
daphne